+++ app/helpers/sub_test_sessions_helper.rb
@@ -1,0 +1,2 @@
+module SubTestSessionsHelper
+end
