+++ app/models/sub_test.rb
@@ -8,7 +8,5 @@ class SubTest < ApplicationRecord
   has_many :subject_sub_tests
   has_many :subjects, through: :subject_sub_tests
   has_many :questions, as: :questionable
-  has_many :answers
-  accepts_nested_attributes_for :answers
   has_one :sub_test_session
 end
