+++ app/models/sub_test_session.rb
@@ -2,4 +2,5 @@ class SubTestSession < ApplicationRecord
   belongs_to :user
   belongs_to :sub_test
   has_many :answers
+  accepts_nested_attributes_for :answers
 end
