+++ db/migrate/20180829112601_create_sub_test_session.rb
@@ -2,7 +2,7 @@ class CreateSubTestSession < ActiveRecord::Migration[5.2]
   def change
     create_table :sub_test_sessions do |t|
       t.references :user
-      t.references :sub_test_sessions
+      t.references :sub_test
     end
   end
 end
