+++ spec/controllers/sub_test_sessions_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe SubTestSessionsController, type: :controller do
+
+end
